backend/requirements.txt
tensorflow-cpu
fastapi
uvicorn
transformers
torch
numpy
pydantic
